+++ pysui/sui/sui_utils.py
@@ -19,6 +19,7 @@ import os
 import base64
 import binascii
 import subprocess
+import json
 import hashlib
 from dataclasses import dataclass
 from pathlib import Path
@@ -65,6 +66,7 @@ from pysui.sui.sui_txresults.single_tx import ObjectRead, ObjectReadData
 _UNPUBLISHED: str = "0000000000000000000000000000000000000000000000000000000000000000"
 
 _SUI_BUILD: list[str] = ["move", "build"]
+_SUI_BUILD2: list[str] = ["move", "build", "--dump", "-e"]
 
 
 @dataclass
@@ -72,6 +74,7 @@ _SUI_BUILD: list[str] = ["move", "build"]
     version="0.66.0",
     reason="Uses raw bytes from modules.",
 )
+@deprecated(version="0.96.0", reason="Changes to new Sui Move package manager.")
 class CompiledPackageRaw:
     """Ease of compilation information dataclass."""
 
@@ -83,6 +86,18 @@ class CompiledPackageRaw:
     package_digest: bytes = None
 
 
+@dataclass
+@versionadded(version="0.96.0", reason="New package manager approach.")
+class CompiledPackage:
+    """Ease of compilation information dataclass."""
+
+    project_source_digest: bytes = None
+    dependencies: list[str] = None
+    compiled_modules: list[bytes] = None
+    package_digest: list[int] = None
+
+
+@deprecated(version="0.96.0", reason="Changes to new Sui Move package manager.")
 def _compile_project(
     path_to_package: Path, args_list: list[str]
 ) -> Union[Path, SuiException]:
@@ -113,6 +128,7 @@ def _compile_project(
     raise SuiPackageBuildFail(result.stdout)
 
 
+@deprecated(version="0.96.0", reason="Changes to new Sui Move package manager.")
 def _compile_projectg(
     sui_bin_str: str, path_to_package: Path, args_list: list[str]
 ) -> Union[Path, SuiException]:
@@ -129,6 +145,30 @@ def _compile_projectg(
     raise SuiPackageBuildFail(result.stdout)
 
 
+@versionadded(version="0.96.0", reason="New package manager approach.")
+def _compile_projectg2(
+    sui_bin_str: str, path_to_package: Path, args_list: list[str], build_env: str
+) -> Union[CompiledPackage, SuiException]:
+    """_compile_projectg Compiles a sui move project in GraohQL environment."""
+    mbs = _SUI_BUILD2.copy()
+    mbs.append(build_env)
+    mbs.extend(args_list)
+    mbs.append("-p")
+    mbs.append(path_to_package)
+
+    mbs.insert(0, sui_bin_str)
+    result = subprocess.run(mbs, capture_output=True, text=True)
+    if result.returncode == 0:
+        jsonb = json.loads(result.stdout)
+        return CompiledPackage(
+            dependencies=jsonb["dependencies"],
+            compiled_modules=[list(base64.b64decode(x)) for x in jsonb["modules"]],
+            project_source_digest=jsonb["digest"],
+        )
+    raise SuiPackageBuildFail(result.stdout)
+
+
+@deprecated(version="0.96.0", reason="Changes to new Sui Move package manager.")
 def _build_dep_info(build_path: str) -> Union[CompiledPackageRaw, Exception]:
     """Fetch details about build."""
     build_info = Path(build_path).joinpath("BuildInfo.yaml")
@@ -152,6 +192,7 @@ def _build_dep_info(build_path: str) -> Union[CompiledPackageRaw, Exception]:
     raise ValueError("Corrupt publish build information")
 
 
+@deprecated(version="0.96.0", reason="Changes to new Sui Move package manager.")
 def _package_digestg(package: CompiledPackageRaw, module_path: Path) -> None:
     """Captures compiled module bytes for publishing and digest calculation."""
 
@@ -177,10 +218,30 @@ def _package_digestg(package: CompiledPackageRaw, module_path: Path) -> None:
     package.compiled_modules = mod_bytes
 
 
+@versionadded(version="0.96.0", reason="New package manager approach.")
+def _package_digestg2(package: CompiledPackage) -> None:
+    """Captures compiled module bytes for publishing and digest calculation."""
+
+    all_digests: list[bytes] = []
+
+    for mmod in package.compiled_modules:
+        hasher = hashlib.blake2b(digest_size=32)
+        hasher.update(mmod[0])
+        all_digests.append(hasher.digest())
+    for dep_str in package.dependencies:
+        all_digests.append(binascii.unhexlify(dep_str[2:]))
+    all_digests.sort()
+    hasher = hashlib.blake2b(digest_size=32)
+    for bblock in all_digests:
+        hasher.update(bblock)
+    package.package_digest = hasher.digest()
+
+
 @versionchanged(
     version="0.17.0",
     reason="Added the package digest that matches chain digest.",
 )
+@deprecated(version="0.96.0", reason="Changes to new Sui Move package manager.")
 def publish_build(
     path_to_package: Path,
     args_list: list[str],
@@ -214,6 +275,7 @@ def publish_build(
     return cpackage
 
 
+@deprecated(version="0.96.0", reason="Changes to new Sui Move package manager.")
 def publish_buildg(
     sui_bin_path_str: str,
     path_to_package: Path,
@@ -243,6 +305,22 @@ def publish_buildg(
     cpackage = _build_dep_info(build_subdir[0].path)
     # Set module bytes as base64 strings and generate package digest
     _package_digestg(cpackage, byte_modules)
+    return cpackage
+
+
+@versionadded(version="0.96.0", reason="New package manager approach.")
+def publish_buildg2(
+    sui_bin_path_str: str, path_to_package: Path, args_list: list[str], build_env: str
+) -> Union[CompiledPackageRaw, Exception]:
+    """Build and collect module base64 strings and dependencies ObjectIDs."""
+    # Compile the package
+    cpackage = _compile_projectg2(
+        sui_bin_path_str, path_to_package, args_list, build_env
+    )
+
+    cpackage.package_digest = bytes(cpackage.project_source_digest)
+
+    # _package_digestg2(cpackage)
     return cpackage